+++ packages/bridge-controller/src/bridge-controller.test.ts
@@ -313,7 +313,10 @@ describe('BridgeController', function () {
       .mockImplementationOnce(async () => {
         return await new Promise((resolve) => {
           return setTimeout(() => {
-            resolve(mockBridgeQuotesNativeErc20Eth as never);
+            resolve({
+              quotes: mockBridgeQuotesNativeErc20Eth as never,
+              validationFailures: [],
+            });
           }, 5000);
         });
       });
@@ -321,10 +324,13 @@ describe('BridgeController', function () {
     fetchBridgeQuotesSpy.mockImplementationOnce(async () => {
       return await new Promise((resolve) => {
         return setTimeout(() => {
-          resolve([
-            ...mockBridgeQuotesNativeErc20Eth,
-            ...mockBridgeQuotesNativeErc20Eth,
-          ] as never);
+          resolve({
+            quotes: [
+              ...mockBridgeQuotesNativeErc20Eth,
+              ...mockBridgeQuotesNativeErc20Eth,
+            ] as never,
+            validationFailures: [],
+          });
         }, 10000);
       });
     });
@@ -340,10 +346,13 @@ describe('BridgeController', function () {
     fetchBridgeQuotesSpy.mockImplementationOnce(async () => {
       return await new Promise((resolve) => {
         return setTimeout(() => {
-          resolve([
-            ...mockBridgeQuotesNativeErc20Eth,
-            ...mockBridgeQuotesNativeErc20Eth,
-          ] as never);
+          resolve({
+            quotes: [
+              ...mockBridgeQuotesNativeErc20Eth,
+              ...mockBridgeQuotesNativeErc20Eth,
+            ] as never,
+            validationFailures: [],
+          });
         }, 10000);
       });
     });
@@ -601,7 +610,10 @@ describe('BridgeController', function () {
       .mockImplementation(async () => {
         return await new Promise((resolve) => {
           return setTimeout(() => {
-            resolve(mockBridgeQuotesSolErc20 as never);
+            resolve({
+              quotes: mockBridgeQuotesSolErc20 as never,
+              validationFailures: [],
+            });
           }, 2000);
         });
       });
@@ -792,18 +804,24 @@ describe('BridgeController', function () {
       .mockImplementationOnce(async () => {
         return await new Promise((resolve) => {
           return setTimeout(() => {
-            resolve(mockBridgeQuotesNativeErc20Eth as never);
+            resolve({
+              quotes: mockBridgeQuotesNativeErc20Eth as never,
+              validationFailures: [],
+            });
           }, 5000);
         });
       });
 
-    fetchBridgeQuotesSpy.mockImplementation(async () => {
+    fetchBridgeQuotesSpy.mockImplementationOnce(async () => {
       return await new Promise((resolve) => {
         return setTimeout(() => {
-          resolve([
-            ...mockBridgeQuotesNativeErc20Eth,
-            ...mockBridgeQuotesNativeErc20Eth,
-          ] as never);
+          resolve({
+            quotes: [
+              ...mockBridgeQuotesNativeErc20Eth,
+              ...mockBridgeQuotesNativeErc20Eth,
+            ] as never,
+            validationFailures: [],
+          });
         }, 10000);
       });
     });
@@ -978,18 +996,24 @@ describe('BridgeController', function () {
       .mockImplementationOnce(async () => {
         return await new Promise((resolve) => {
           return setTimeout(() => {
-            resolve(mockBridgeQuotesNativeErc20Eth as never);
+            resolve({
+              quotes: mockBridgeQuotesNativeErc20Eth as never,
+              validationFailures: [],
+            });
           }, 5000);
         });
       });
 
-    fetchBridgeQuotesSpy.mockImplementation(async () => {
+    fetchBridgeQuotesSpy.mockImplementationOnce(async () => {
       return await new Promise((resolve) => {
         return setTimeout(() => {
-          resolve([
-            ...mockBridgeQuotesNativeErc20Eth,
-            ...mockBridgeQuotesNativeErc20Eth,
-          ] as never);
+          resolve({
+            quotes: [
+              ...mockBridgeQuotesNativeErc20Eth,
+              ...mockBridgeQuotesNativeErc20Eth,
+            ] as never,
+            validationFailures: [],
+          });
         }, 10000);
       });
     });
@@ -1256,7 +1280,10 @@ describe('BridgeController', function () {
         .mockImplementationOnce(async () => {
           return await new Promise((resolve) => {
             return setTimeout(() => {
-              resolve(quoteResponse as never);
+              resolve({
+                quotes: quoteResponse as never,
+                validationFailures: [],
+              });
             }, 1000);
           });
         });
@@ -1383,7 +1410,10 @@ describe('BridgeController', function () {
     fetchBridgeQuotesSpy.mockImplementationOnce(async () => {
       return await new Promise((resolve) => {
         return setTimeout(() => {
-          resolve(mockBridgeQuotesNativeErc20Eth as never);
+          resolve({
+            quotes: mockBridgeQuotesNativeErc20Eth,
+            validationFailures: [],
+          } as never);
         }, 1000);
       });
     });
@@ -1468,6 +1498,7 @@ describe('BridgeController', function () {
     [
       'should append solanaFees for Solana quotes',
       mockBridgeQuotesSolErc20 as unknown as QuoteResponse[],
+      [],
       2,
       '5000',
       '300',
@@ -1475,6 +1506,7 @@ describe('BridgeController', function () {
     [
       'should not append solanaFees if selected account is not a snap',
       mockBridgeQuotesSolErc20 as unknown as QuoteResponse[],
+      [],
       2,
       undefined,
       '0',
@@ -1486,6 +1518,22 @@ describe('BridgeController', function () {
         ...mockBridgeQuotesSolErc20,
         ...mockBridgeQuotesErc20Native,
       ] as unknown as QuoteResponse[],
+      [],
+      8,
+      undefined,
+      '1',
+    ],
+    [
+      'should handle malformed quotes',
+      [
+        ...mockBridgeQuotesSolErc20,
+        ...mockBridgeQuotesErc20Native,
+      ] as unknown as QuoteResponse[],
+      [
+        'socket|quote.srcAsset.decimals',
+        'socket|quote.destAsset.address',
+        'lifi|quote.srcAsset.decimals',
+      ],
       8,
       undefined,
       '1',
@@ -1495,6 +1543,7 @@ describe('BridgeController', function () {
     async (
       _testTitle: string,
       quoteResponse: QuoteResponse[],
+      validationFailures: string[],
       expectedQuotesLength: number,
       expectedFees: string | undefined,
       expectedMinBalance: string | undefined,
@@ -1585,7 +1634,10 @@ describe('BridgeController', function () {
         .mockImplementation(async () => {
           return await new Promise((resolve) => {
             return setTimeout(() => {
-              resolve(quoteResponse as never);
+              resolve({
+                quotes: quoteResponse,
+                validationFailures,
+              });
             }, 1000);
           });
         });
@@ -1653,6 +1705,17 @@ describe('BridgeController', function () {
       expect(snapCalls).toMatchSnapshot();
 
       expect(quotes).toHaveLength(expectedQuotesLength);
+
+      // Verify validation failure tracking
+      expect(trackMetaMetricsFn).toHaveBeenCalledTimes(
+        6 + (validationFailures.length ? 1 : 0),
+      );
+      expect(
+        trackMetaMetricsFn.mock.calls.filter(
+          ([eventName]) =>
+            eventName === UnifiedSwapBridgeEventName.QuotesValidationFailed,
+        ),
+      ).toMatchSnapshot();
     },
   );
 
@@ -1991,6 +2054,37 @@ describe('BridgeController', function () {
 
       expect(trackMetaMetricsFn.mock.calls).toMatchSnapshot();
     });
+
+    it('should track the StatusValidationFailed event', () => {
+      const controller = new BridgeController({
+        messenger: messengerMock,
+        getLayer1GasFee: getLayer1GasFeeMock,
+        clientId: BridgeClientId.EXTENSION,
+        fetchFn: mockFetchFn,
+        trackMetaMetricsFn,
+        state: {
+          quoteRequest: {
+            srcChainId: SolScope.Mainnet,
+            destChainId: '1',
+            srcTokenAddress: 'NATIVE',
+            destTokenAddress: '0x1234',
+            srcTokenAmount: '1000000',
+            walletAddress: '0x123',
+            slippage: 0.5,
+          },
+          quotes: mockBridgeQuotesSolErc20 as never,
+        },
+      });
+      controller.trackUnifiedSwapBridgeEvent(
+        UnifiedSwapBridgeEventName.StatusValidationFailed,
+        {
+          failures: ['Failed to submit tx'],
+        },
+      );
+      expect(trackMetaMetricsFn).toHaveBeenCalledTimes(1);
+
+      expect(trackMetaMetricsFn.mock.calls).toMatchSnapshot();
+    });
   });
 
   describe('trackUnifiedSwapBridgeEvent client-side call exceptions', () => {
@@ -2085,6 +2179,7 @@ describe('BridgeController', function () {
     quotesByDecreasingProcessingTime.reverse();
 
     beforeEach(() => {
+      jest.clearAllMocks();
       jest
         .spyOn(featureFlagUtils, 'getBridgeFeatureFlags')
         .mockReturnValueOnce({
@@ -2102,7 +2197,10 @@ describe('BridgeController', function () {
     it('should override aggIds and noFee in perps request', async () => {
       const fetchBridgeQuotesSpy = jest
         .spyOn(fetchUtils, 'fetchBridgeQuotes')
-        .mockResolvedValueOnce(quotesByDecreasingProcessingTime as never);
+        .mockResolvedValueOnce({
+          quotes: quotesByDecreasingProcessingTime as never,
+          validationFailures: [],
+        });
       const expectedControllerState = bridgeController.state;
 
       const quotes = await bridgeController.fetchQuotes(
@@ -2162,7 +2260,10 @@ describe('BridgeController', function () {
     it('should add aggIds and noFee to perps request', async () => {
       const fetchBridgeQuotesSpy = jest
         .spyOn(fetchUtils, 'fetchBridgeQuotes')
-        .mockResolvedValueOnce(quotesByDecreasingProcessingTime as never);
+        .mockResolvedValueOnce({
+          quotes: quotesByDecreasingProcessingTime as never,
+          validationFailures: [],
+        });
       const expectedControllerState = bridgeController.state;
 
       const quotes = await bridgeController.fetchQuotes(
@@ -2219,7 +2320,10 @@ describe('BridgeController', function () {
     it('should not add aggIds and noFee if featureId is not specified', async () => {
       const fetchBridgeQuotesSpy = jest
         .spyOn(fetchUtils, 'fetchBridgeQuotes')
-        .mockResolvedValueOnce(mockBridgeQuotesSolErc20 as never);
+        .mockResolvedValueOnce({
+          quotes: mockBridgeQuotesSolErc20 as never,
+          validationFailures: [],
+        });
       const expectedControllerState = bridgeController.state;
 
       const quotes = await bridgeController.fetchQuotes(