+++ packages/sample-controllers/src/sample-gas-prices-service/index.ts
@@ -1,2 +1,0 @@
-export { type SampleAbstractGasPricesService } from './sample-abstract-gas-prices-service';
-export { SampleGasPricesService } from './sample-gas-prices-service';